Move & update import to all views
