+++ src/views/ArticleView.vue
@@ -173,17 +173,14 @@
 </template>
 
 <script>
-import BtnElt from "../../components/elements/BtnElt"
-import CardElt from "../../components/elements/CardElt"
-import FieldElt from "../../components/elements/FieldElt"
-import ListElt from "../../components/elements/ListElt"
-import MediaElt from "../../components/elements/MediaElt"
-
-import { checkRange, checkRegex, checkRole } from "../../services/checkers"
-import { deleteData, getData, putData } from "../../services/fetchers"
-import { setError, setMeta } from "../../services/setters"
-
+import BtnElt from "../components/BtnElt"
+import CardElt from "../components/CardElt"
 import Editor from "@tinymce/tinymce-vue"
+import FieldElt from "../components/FieldElt"
+import ListElt from "../components/ListElt"
+import MediaElt from "../components/MediaElt"
+
+import { checkRange, checkRegex, checkRole, deleteData, getData, putData, setError, setMeta } from "../app/services"
 import { mapState } from "vuex"
 
 export default {
