+++ src/views/LinkView.vue
@@ -60,16 +60,13 @@
 </template>
 
 <script>
-import BtnElt from "../../components/elements/BtnElt"
-import CardElt from "../../components/elements/CardElt"
-import ListElt from "../../components/elements/ListElt"
-import NavElt from "../../components/elements/NavElt"
-import LinkSet from "../../components/setters/LinkSet"
+import BtnElt from "../components/BtnElt"
+import CardElt from "../components/CardElt"
+import ListElt from "../components/ListElt"
+import LinkSet from "../components/LinkSet"
+import NavElt from "../components/NavElt"
 
-import { checkRole } from "../../services/checkers"
-import { getCats, getItemsByCat } from "../../services/getters"
-import { setMeta } from "../../services/setters"
-
+import { checkRole, getCats, getItemsByCat, setMeta } from "../app/services"
 import { mapState, mapActions } from "vuex"
 
 export default {
