+++ src/views/PortfolioView.vue
@@ -79,16 +79,13 @@
 </template>
 
 <script>
-import CardElt from "@/components/elements/CardElt"
-import ListElt from "@/components/elements/ListElt"
-import MediaElt from "@/components/elements/MediaElt"
-import NavElt from "@/components/elements/NavElt"
-import ProjectSet from "@/components/setters/ProjectSet"
+import CardElt from "../components/CardElt"
+import ListElt from "../components/ListElt"
+import MediaElt from "../components/MediaElt"
+import NavElt from "../components/NavElt"
+import ProjectSet from "../components/ProjectSet"
 
-import { checkRole } from "@/services/checkers"
-import { getCats, getItemsByCat } from "@/services/getters"
-import { setMeta } from "@/services/setters"
-
+import { checkRole, getCats, getItemsByCat, setMeta } from "../app/services"
 import { mapState, mapActions } from "vuex"
 
 export default {
